+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Changes
 
+- Changed "ascii" to "ansi", woops.
+
 #### Removals
 
 #### Fixes